Remove all source and API not tied to fetched results controller and entity monitor
