+++ .ruby-version
@@ -1,1 +1,1 @@
-2.2.5
+2.4.3
