+++ CoreDataStack.xcodeproj/xcshareddata/xcschemes/CoreDataStack.xcscheme
@@ -35,7 +35,7 @@
             </BuildableReference>
          </BuildActionEntry>
          <BuildActionEntry
-            buildForTesting = "YES"
+            buildForTesting = "NO"
             buildForRunning = "NO"
             buildForProfiling = "NO"
             buildForArchiving = "NO"
@@ -63,16 +63,6 @@
                BlueprintIdentifier = "E4C9BC371AEA848600A6BD1B"
                BuildableName = "CoreDataStackTests.xctest"
                BlueprintName = "CoreDataStackTests"
-               ReferencedContainer = "container:CoreDataStack.xcodeproj">
-            </BuildableReference>
-         </TestableReference>
-         <TestableReference
-            skipped = "NO">
-            <BuildableReference
-               BuildableIdentifier = "primary"
-               BlueprintIdentifier = "E435995B1D932458003295AE"
-               BuildableName = "PostDeprecationTests.xctest"
-               BlueprintName = "PostDeprecationTests"
                ReferencedContainer = "container:CoreDataStack.xcodeproj">
             </BuildableReference>
          </TestableReference>
