+++ Gemfile
@@ -2,4 +2,4 @@ source "https://rubygems.org"
 
 gem "fastlane"
 gem 'jazzy'
-gem 'cocoapods', :git => "git://github.com/CocoaPods/CocoaPods.git", :branch => "master"
+gem 'cocoapods'
