--- Sample Model/Container/Book+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  Book+CoreDataClass.swift
-//  CoreDataStack
-//
-//  Created by Robert Edwards on 9/27/16.
-//  Copyright © 2016 Big Nerd Ranch. All rights reserved.
-//
-
-import Foundation
-import CoreData
-
-@objc(Book)
-public class Book: NSManagedObject {
-
-}
