--- Sample Model/Container/Publisher+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  Publisher+CoreDataClass.swift
-//  CoreDataStack
-//
-//  Created by Robert Edwards on 9/27/16.
-//  Copyright © 2016 Big Nerd Ranch. All rights reserved.
-//
-
-import Foundation
-import CoreData
-
-@objc(Publisher)
-public class Publisher: NSManagedObject {
-
-}
